pyproject.toml
[project]
name = "desktop-moodbot"
version = "0.1.0"
description = "AI agent mood display for your desk"
requires-python = ">=3.10"
license = "LicenseRef-NonCommercial"
dependencies = [
    "vaderSentiment>=3.3.2",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
